Big refactor of the server structure
