+++ server/plaid/index.js
@@ -1,0 +1,23 @@
+const plaid = require("plaid");
+const dotenv = require("dotenv");
+
+const result = dotenv.config();
+if (result.error) {
+  throw result.error;
+}
+const PLAID_CLIENT_ID = process.env.PLAID_CLIENT_ID;
+const PLAID_SECRET = process.env.PLAID_SECRET;
+const PLAID_PUBLIC_KEY = process.env.PLAID_PUBLIC_KEY;
+const PLAID_ENV = process.env.PLAID_ENV;
+
+// Initialize the Plaid client
+// Find your API keys in the Dashboard (https://dashboard.plaid.com/account/keys)
+const client = new plaid.Client(
+  PLAID_CLIENT_ID,
+  PLAID_SECRET,
+  PLAID_PUBLIC_KEY,
+  plaid.environments[PLAID_ENV],
+  { version: "2018-05-22" }
+);
+
+module.exports = client;
