+++ server/routes/pocketgoblin.routes.js
@@ -1,0 +1,4 @@
+const ctrl = require("../controllers/pocketgoblin.controller.js");
+module.exports = function(app) {
+  app.get("/transactions", ctrl.getTransactions);
+};
